wip: fps counter

+++ Timer.h
@@ -5,7 +5,7 @@ namespace Timer
 	extern unsigned int FRAME_COUNT;
 	extern float DELTA_TIME;
 	extern float FPS_RAW;
-	extern float FPS_STABLE;
+	extern unsigned int FPS_STABLE;
 
 	void Init();
 	void NewFrame();
